Add git branch to `$PS1` in `dotfiles/home/.bashrc`
It's nice to see which branch you're in at-a-glance. This change makes it
so that, if you are in a git repo, the `$PS1` displays which branch is
active. If not, it is reduced to a single space. I would like to find a
way to conditionally add a space between (BRANCH) and DIR, omitting it if
not in a git repo. This would make it look nicer outside of git repos.